refactor: slim down the upload handler
