+++ server/routes/upload/helpers/cleanDeckname.ts
@@ -1,0 +1,8 @@
+export default function cleanDeckName(name: string) {
+  let _name = name;
+  if (name.startsWith("&#x")) {
+    _name = name.split(" ").slice(1).join("").trim();
+  }
+  return _name;
+}
+
